feat: refactor logo request logging to UserService .
